Moving live services to services/live/ dir.
