+++ src/lib/io/live-downloader.ts
@@ -1,7 +1,7 @@
-import { LiveTldListerService } from '../services/domains.js';
-import { LiveEmailProviderListerService } from '../services/email-providers.js';
-import LiveHubspotService from '../services/hubspot.js';
-import { LiveMarketplaceService } from '../services/marketplace.js';
+import { LiveTldListerService } from '../services/live/domains.js';
+import { LiveEmailProviderListerService } from '../services/live/email-providers.js';
+import LiveHubspotService from '../services/live/hubspot.js';
+import { LiveMarketplaceService } from '../services/live/marketplace.js';
 import { Downloader } from './interfaces.js';
